+++ src/Query/LoginQuery.php
@@ -17,40 +17,16 @@ use Studoo\Api\EcoleDirecte\Entity\Login;
  * Traitement de la requête de connexion par requête API EcoleDirecte
  * @package Studoo\Api\EcoleDirecte\Query
  */
-class LoginQuery implements EntityQueryInterface
+class LoginQuery extends Query implements EntityQueryInterface
 {
-    private string $methode = "POST";
-    private string $path = "login.awp";
-    private array $query = [
-        'identifiant' => '',
-        'motdepasse' => ''
-    ];
-
-    /**
-     * Retourne la méthode de la requête
-     * @return string
-     */
-    public function getMethode(): string
+    public function __construct()
     {
-        return $this->methode;
-    }
-
-    /**
-     * Retourne le chemin de la requête
-     * @return string
-     */
-    public function getPath(): string
-    {
-        return $this->path;
-    }
-
-    /**
-     * retourne les paramètres de la requête
-     * @return array
-     */
-    public function getQuery(): array
-    {
-        return $this->query;
+        $this->methode = 'POST';
+        $this->path = 'login.awp';
+        $this->query = [
+            'identifiant' => '',
+            'motdepasse' => ''
+        ];
     }
 
     /**
